@@ -6,10 +6,11 @@ description: >-
   A chrome extension for Dart debugging.
 
 environment:
-  sdk: ">=2.5.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   async: ^2.3.0
+  collection: ^1.15.0
   js: ^0.6.1+1
   pub_semver: ^2.0.0
   sse: ^4.1.0
